refactored signature suite
