+++ core/src/proof.rs
@@ -14,9 +14,9 @@ pub struct DataIntegrityProof {
 /// Given a JSON-LD document, create a data integrity proof for the document.
 /// Currently, only `Ed25519Signature2018` data integrity proofs in the JSON-LD format can be created.
 pub fn create_data_integrity_proof<S: signature::Signature>(
-    signer: &impl signature::DIDSigner<S>,
+    signer: &impl signature::signer::DIDSigner<S>,
     doc: serde_json::Value,
-    relation: signature::VerificationRelation,
+    relation: signature::suite::VerificationRelation,
 ) -> Result<DataIntegrityProof, Box<dyn std::error::Error>> {
     let mut hasher = Sha512::new();
     hasher.update(normalization::normalize(doc));
@@ -38,16 +38,17 @@ mod tests {
     use sha2::Digest;
 
     use super::create_data_integrity_proof;
-    use signature::{DIDSigner, DIDVerifier};
+    use signature::signer::DIDSigner;
+    use signature::verifier::DIDVerifier;
 
     #[rstest::rstest]
-    #[case::success(json!("{}"), signature::VerificationRelation::AssertionMethod)]
+    #[case::success(json!("{}"), signature::suite::VerificationRelation::AssertionMethod)]
     fn test_create_data_integrity_proof(
         #[case] doc: serde_json::Value,
-        #[case] relation: signature::VerificationRelation,
+        #[case] relation: signature::suite::VerificationRelation,
     ) {
-        let signer = signature::Ed25519DidSigner::new();
-        let verifier = signature::Ed25519DidVerifier::from(&signer);
+        let signer = signature::signer::Ed25519DidSigner::new();
+        let verifier = signature::verifier::Ed25519DidVerifier::from(&signer);
         let res = create_data_integrity_proof(&signer, doc.clone(), relation);
 
         assert!(res.is_ok());
